refactor(alpha): trim stale module surfaces
