+++ docs/content/1.getting-started/2.configuration.md
@@ -60,22 +60,6 @@ export default defineNuxtConfig({
     **Requirement**: Requires [NuxtHub Integration](/integrations/nuxthub) with `hub: { kv: true }`.
   ::
 
-  ::field{name="database.provider" type="'none' | 'nuxthub' | 'convex'"}
-    Default: auto (`'nuxthub'` when `hub.db` is configured, otherwise `'none'`)
-
-    Select the database backend used by Better Auth.
-    - `'none'`: No DB adapter
-    - `'nuxthub'`: Use `@nuxthub/core` database
-    - `'convex'`: Use Convex HTTP adapter (requires `nuxt-convex`)
-  ::
-
-  ::field{name="database.convexUrl" type="string"}
-    Default: auto-detected
-
-    Optional Convex URL override when `database.provider = 'convex'`.
-    Priority: `auth.database.convexUrl` → `convex.url` → `runtimeConfig.public.convex.url` → `CONVEX_URL` → `NUXT_PUBLIC_CONVEX_URL`
-  ::
-
   ::field{name="schema.usePlural" type="boolean"}
     Default: `false`
 
@@ -129,7 +113,7 @@ export default defineServerAuth((ctx) => ({
 
   // Access the database connection via ctx.db when using NuxtHub
   // Example: pass ctx.db to your own plugin/helper.
-  // Do not set `database` here - the module injects it from `auth.database.provider`.
+  // Do not set `database` here when using module-managed adapters.
   appName: ctx.runtimeConfig.public.siteUrl ? 'Better Auth App' : 'Better Auth',
 
   // Access runtime config if needed
