+++ docs/content/4.integrations/3.convex.md
@@ -1,111 +1,19 @@
 ---
 title: Convex
-description: Use Convex as your real-time backend with Better Auth.
+description: Current status of Convex integration in this module.
 ---
 
-Convex provides a real-time backend with automatic data synchronization. You can combine it with Better Auth to build applications where authenticated users access their own data with live updates.
+This module does not currently provide a built-in Convex adapter integration.
 
-## Setup
+## Status
 
-The `nuxt-convex` module connects your Nuxt application to Convex. Better Auth handles authentication while Convex manages your application data.
+- Convex-specific module configuration (`auth.database.provider = 'convex'`) is not supported.
+- The maintained paths in this module are NuxtHub-backed storage and database-less/client-only usage.
 
-::steps
+## Recommended paths
 
-### Install nuxt-convex
+- For external Better Auth backends, use [External Auth Backend](/guides/external-auth-backend).
+- For Convex-specific auth/database setup, use the upstream Convex and Better Auth documentation.
 
-```bash
-npx nuxt module add nuxt-convex
-```
-
-### Configure modules
-
-Add both modules to your Nuxt configuration.
-
-```ts [nuxt.config.ts]
-export default defineNuxtConfig({
-  modules: ['nuxt-convex', '@onmax/nuxt-better-auth'],
-  auth: {
-    database: {
-      provider: 'convex',
-      // optional override; otherwise auto-resolved from convex config/env
-      // convexUrl: 'https://your-app.convex.cloud'
-    },
-  },
-})
-```
-
-::
-
-::callout{icon="i-lucide-book" to="https://nuxt-convex.onmax.me" target="_blank"}
-See the [nuxt-convex documentation](https://nuxt-convex.onmax.me) for Convex project setup and environment configuration.
-::
-
-## Authenticated Queries
-
-Better Auth provides the authenticated user through the `useUserSession` composable. You can pass the user ID to Convex queries to fetch user-specific data with real-time updates.
-
-```vue [app/pages/tasks.vue]
-<script setup lang="ts">
-import { api } from '#convex/api'
-
-const { user } = useUserSession()
-const { data: tasks } = useConvexQuery(
-  api.tasks.list,
-  computed(() => user.value ? { userId: user.value.id } : undefined),
-  { ssr: false }
-)
-</script>
-```
-
-The `computed` wrapper ensures the query re-subscribes automatically when the user changes. When you pass `undefined` as the arguments, Convex pauses the query and returns no data. This behavior prevents unauthorized queries when the user is not authenticated.
-
-::warning
-Authenticated queries require `{ ssr: false }`. Convex queries executed on the server during SSR do not receive authentication cookies, which causes them to fail or return empty results.
-::
-
-## Convex as Auth Database
-
-Better Auth typically stores authentication data in a SQL database. The `@convex-dev/better-auth` adapter lets you store users, sessions, and accounts directly in Convex instead. This approach keeps all your data in one place with Convex's real-time synchronization.
-
-### Install the adapter
-
-The adapter requires exact versions to ensure compatibility between Better Auth and Convex.
-
-```bash
-pnpm add @convex-dev/better-auth better-auth --save-exact
-```
-
-### Configure the Convex component
-
-The Better Auth component registers the authentication tables in your Convex database.
-
-```ts [convex/convex.config.ts]
-import betterAuth from '@convex-dev/better-auth/convex.config'
-import { defineApp } from 'convex/server'
-
-const app = defineApp()
-app.use(betterAuth)
-export default app
-```
-
-### Connect Better Auth to Convex
-
-Set `auth.database.provider = 'convex'` in `nuxt.config.ts`. The module injects the Convex HTTP adapter automatically. In `server/auth.config.ts`, keep your auth options and add the Convex plugin.
-
-```ts [server/auth.config.ts]
-import { convex } from '@convex-dev/better-auth/server/plugins'
-import { defineServerAuth } from '@onmax/nuxt-better-auth/config'
-
-export default defineServerAuth({
-  plugins: [convex()],
-  emailAndPassword: { enabled: true },
-})
-```
-
-::note
-Convex auth tables are generated at `.nuxt/better-auth/auth-tables.convex.ts`.
-::
-
-::callout{icon="i-lucide-book" to="https://github.com/get-convex/convex-better-auth" target="_blank"}
-See the [@convex-dev/better-auth repository](https://github.com/get-convex/convex-better-auth) for advanced configuration options and additional plugins.
-::
+:read-more{to="https://www.convex.dev/docs"}
+:read-more{to="https://www.better-auth.com/docs"}
