+++ docs/content/6.troubleshooting/1.faq.md
@@ -1,80 +1,19 @@
 ---
 title: Troubleshooting
-description: Common issues and their solutions.
+description: Common issues and fast checks for current module features.
 ---
 
-## Debugging
+## Quick checks
 
-Enable verbose logging:
+- `BETTER_AUTH_SECRET` is set and at least 32 characters in production.
+- `runtimeConfig.public.siteUrl` (or `NUXT_PUBLIC_SITE_URL`) is set when callback consistency matters.
+- `server/auth.config.ts` and `app/auth.config.ts` both export their `define*Auth(...)` default config.
 
-```ini [.env]
-DEBUG=better-auth*
-```
+## Common issues
 
-Check browser DevTools Network tab for `/api/auth/*` requests.
+### Invalid origin or OAuth callback mismatch
 
-## Frequently Asked Questions
-
-### Can I use this without a database?
-
-Yes, using [database-less mode](/guides/database-less-mode). Sessions are stored in encrypted cookies. This mode limits some features like session revocation.
-
-### Does this work with Nuxt 2?
-
-No. This module requires Nuxt 4 or later.
-
-### What's the difference from nuxt-auth-utils?
-
-Nuxt Better Auth uses the Better Auth library, which provides:
-- More authentication methods (OAuth, magic links, passkeys)
-- Plugin ecosystem for features like 2FA, organizations
-- TypeScript-first design with full type inference
-
-See the [migration guide](/guides/migrate-from-nuxt-auth-utils) for detailed comparison.
-
-### Can I use a different database?
-
-Yes. Better Auth supports any database with a Drizzle, Prisma, or Kysely adapter. See [Better Auth database documentation](https://www.better-auth.com/docs/database).
-
-### How do I add custom fields to the user?
-
-Configure in Better Auth, then augment types:
-
-```ts [server/auth.config.ts]
-import { defineServerAuth } from '@onmax/nuxt-better-auth/config'
-
-export default defineServerAuth({
-  user: {
-    additionalFields: {
-      role: { type: 'string', defaultValue: 'user' }
-    }
-  }
-})
-```
-
-:read-more{to="https://www.better-auth.com/docs/concepts/users-accounts#additional-fields"}
-
-### How do I enable email verification?
-
-Use Better Auth's email plugin:
-
-:read-more{to="https://www.better-auth.com/docs/plugins/email-otp"}
-
-### How do I implement password reset?
-
-:read-more{to="https://www.better-auth.com/docs/authentication/email-password#forget-password"}
-
-## Common Errors
-
-## "Invalid origin" locally (followed by provider "invalid redirect_uri")
-
-This error usually means Better Auth resolves a different local host than your browser origin. The most common mismatch is `127.0.0.1` vs `localhost`.
-
-Follow this workflow to debug the mismatch:
-
-::steps
-### Set an explicit local site URL
-Add an explicit site URL in `nuxt.config.ts`:
+Set an explicit local site URL and restart dev:
 
 ```ts [nuxt.config.ts]
 export default defineNuxtConfig({
@@ -86,60 +25,32 @@ export default defineNuxtConfig({
 })
 ```
 
-### Restart and inspect auth config
-Restart dev and inspect [http://localhost:3000/api/_better-auth/config](http://localhost:3000/api/_better-auth/config):
+Then verify your provider callback URL matches exactly.
 
-- Verify `config.server.baseURL` matches your local origin exactly.
-- Verify `config.server.trustedOrigins` includes your local origin.
+### Missing auth config files
 
-When `runtimeConfig.public.siteUrl` points to production during local development, the module still adds local dev origins to `trustedOrigins` (request origin plus localhost fallback). This supports `--host` and LAN testing as long as your request origin is stable.
-::
+Create these files if missing:
 
-::important
-OAuth providers must use exact callback URL matching. Ensure provider settings match your app callback exactly (including protocol, host, port, and path), for example:
+- `server/auth.config.ts` with `export default defineServerAuth(...)`
+- `app/auth.config.ts` with `export default defineClientAuth(...)`
 
-- `http://localhost:3000/api/auth/callback/github`
-::
+### No schema file is generated
 
-## NuxtHub version requirement
+Schema generation requires NuxtHub with DB enabled:
 
-If using NuxtHub for database features, you need `@nuxthub/core` version **0.10.5** or higher. NuxtHub is optional - see [NuxtHub Integration](/integrations/nuxthub) for details.
-
-```bash
-pnpm add @nuxthub/core@latest
+```ts [nuxt.config.ts]
+export default defineNuxtConfig({
+  hub: { db: { dialect: 'sqlite' } },
+})
 ```
 
-## "Missing server/auth.config.ts"
+### API routes are not protected
 
-The module validates config files at build time.
-Create `server/auth.config.ts` exporting `defineServerAuth(...)`.
+Only routes with `routeRules.auth` are enforced automatically for `/api/**`.
+Use `routeRules.auth` or call `requireUserSession(event)` inside handlers.
 
-## "Missing <srcDir>/auth.config.ts"
+## Related docs
 
-The module expects a client config file at `<srcDir>/auth.config.ts` exporting `defineClientAuth(...)` as default.
-
-## "auth.config.ts validation failed"
-
-- Ensure the file exports `defineServerAuth(...)`.
-- Do not set `secret`/`baseURL` manually (the module injects these).
-- Confirm the file is at the expected path (default: `server/auth.config.ts`).
-
-## No schema file is generated
-
-Schema generation requires NuxtHub. Ensure NuxtHub is installed and DB is configured:
-
-```ts
-hub: { db: { dialect: 'sqlite' } }
-```
-
-Restart dev server after changing Better Auth plugins.
-
-## `routeRules` changes don't apply
-
-`routeRules` are synced to page meta during `pages:extend`.
-Restart dev server after updating route rules.
-
-## API routes aren't protected
-
-Only `routeRules.auth` is enforced for `/api/**`.
-Use `routeRules.auth` (or call `requireUserSession(event)` inside handlers) to enforce access.
+- [Configuration](/getting-started/configuration)
+- [Route Protection](/core-concepts/route-protection)
+- [NuxtHub Integration](/integrations/nuxthub)
